% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManipulatePDFs.R
\name{combine_pdfs}
\alias{combine_pdfs}
\title{combine pdfs}
\usage{
combine_pdfs(files, output_name = "output", dir = NULL)
}
\arguments{
\item{files}{files}

\item{output_name}{output.pdf by default; only specify name, not extension}

\item{dir}{NULL by default so will put in common path of the files}
}
\description{
combine pdfs
}
