% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DevelopPackages.R
\name{copy_files_over}
\alias{copy_files_over}
\title{Copy files over for creating new packages}
\usage{
copy_files_over(...)
}
\arguments{
\item{...}{filepaths of the files}
}
\value{
list of files copied
}
\description{
Copy files over for creating new packages
}
\examples{
if(FALSE) {
copy_files_over(
fs::path_home("OneDrive/R Functions/utilities.R"),
fs::path_home("OneDrive/R Functions/irt.R")
)}

}
