% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DevelopPackages.R
\name{DevelopMyPackage}
\alias{DevelopMyPackage}
\title{Create a package just how I like it <3}
\usage{
DevelopMyPackage(
  package_name,
  rstudio = rstudioapi::isAvailable(),
  open = rlang::is_interactive(),
  roxygen = TRUE,
  check_name = TRUE,
  use_email = c("personal", "work", "city")
)
}
\arguments{
\item{package_name}{Package name to use. By default, uses \code{basename(path)}.
If \code{path == '.'} & \code{package_name} is not explicitly set,
then \code{basename(getwd())} will be used.}

\item{rstudio}{Boolean. In RStudio right now?}

\item{open}{Boolean. Open the created project?}

\item{roxygen}{Boolean. Use roxygen?}

\item{check_name}{Should we check that the package name is correct according to CRAN requirements.}
}
\value{
The path, invisibly.
}
\description{
Create a package just how I like it <3
}
\note{
checke golem::create_golem for more inspiration;
magrittr pipe
example DevelopMyPackage("bunny")
}
