% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindingFunctions.R
\name{extract_from_helpfiles}
\alias{extract_from_helpfiles}
\title{extract_from_helpfiles
See C:/Users/Natalia/OneDrive/PhD Psychology/01 - R Project/finding-func.R}
\usage{
extract_from_helpfiles(helpfile, selector = "p")
}
\arguments{
\item{helpfile}{helpfile}

\item{selector}{"p"}
}
\description{
extract_from_helpfiles
See C:/Users/Natalia/OneDrive/PhD Psychology/01 - R Project/finding-func.R
}
