% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindingFunctions.R
\name{fetch_helpfiles}
\alias{fetch_helpfiles}
\title{fetch_helpfiles
See C:/Users/Natalia/OneDrive/PhD Psychology/01 - R Project/finding-func.R}
\usage{
fetch_helpfiles(func_name, pkg = "base")
}
\arguments{
\item{func_name}{function name}

\item{pkg}{package name e.g., "base"}
}
\description{
fetch_helpfiles
See C:/Users/Natalia/OneDrive/PhD Psychology/01 - R Project/finding-func.R
}
