% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HappyFeedback.R
\name{HappyFeedback}
\alias{HappyFeedback}
\title{Lighten Natalia's mood}
\usage{
HappyFeedback()
}
\value{
a happy feedback
}
\description{
Primarily a wrapper around bhappyr to make the output colourful!
TODO: Build on the function more to introduce more sentence possibilities + bunnies
}
\examples{
library(etc)
HappyFeedback()

}
