% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindingFunctions.R
\name{lsp}
\alias{lsp}
\title{lsp
See C:/Users/Natalia/OneDrive/PhD Psychology/01 - R Project/finding-func.R}
\usage{
lsp(package, what, pattern)
}
\arguments{
\item{package}{pkg}

\item{what}{what}

\item{pattern}{pattern}
}
\description{
lsp
See C:/Users/Natalia/OneDrive/PhD Psychology/01 - R Project/finding-func.R
}
