% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuplicatesManager.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{remove_duplicates}
\usage{
remove_duplicates(df, bad_dir, auto_remove = FALSE)
}
\arguments{
\item{df}{output from identify_duplicates()}

\item{bad_dir}{directories that shouldn't contain the files}

\item{auto_remove}{TRUE or FALSE; default is FALSE}
}
\value{
a list
}
\description{
remove_duplicates
}
