% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManipulatePDFs.R
\name{rotate_pdf}
\alias{rotate_pdf}
\title{rotate_pdf}
\usage{
rotate_pdf(file_name, dir, rotation_degrees = 90)
}
\arguments{
\item{file_name}{e.g., "NM.pdf"}

\item{dir}{e.g., fs::path_home_r("01 - Important Documents/NM/BRP")}

\item{rotation_degrees}{numeric, default 90}
}
\description{
rotate_pdf
}
