% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchiveManager.R
\name{sort_my_zips}
\alias{sort_my_zips}
\title{sort_my_zips}
\usage{
sort_my_zips(all = FALSE, recurse = FALSE)
}
\arguments{
\item{all}{all posible archive formats or just .zip?}

\item{recurse}{used in {fs} pkg functions. If TRUE, recurse fully.
If a positive number, the number of levels to recurse.}
}
\value{
a list
}
\description{
sort_my_zips
}
\examples{
\dontrun{
out <- sort_my_zips()
out$zips_to_check #fix these
out$problematic_extractions$name #check these
fs::file_delete(out$my_zips$filename) #clean Up
}

}
