% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManipulatePDFs.R
\name{split_pdf}
\alias{split_pdf}
\title{split pdf}
\usage{
split_pdf(file_name, dir, pages)
}
\arguments{
\item{dir}{dir}

\item{pages}{e.g., 1 or 1:3 or 3:5 etc}

\item{file}{file}
}
\description{
create a new pdf with a subset of the input pages
A wrapper around \code{\link[qpdf:qpdf]{qpdf::pdf_subset()}}
}
