% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc-package.R
\docType{package}
\name{etc-package}
\alias{etc}
\alias{etc-package}
\title{etc: Extraneous Functions for Natalia's Benefit}
\description{
see above.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/natalia-mlad/etc}
  \item Report bugs at \url{https://github.com/natalia-mlad/etc/issues}
}

}
\author{
\strong{Maintainer}: Natalia Mladentseva \email{natashka.ml@gmail.com}

}
\keyword{internal}
