% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchiveManager.R
\name{extract_my_zips}
\alias{extract_my_zips}
\title{extract_my_zips}
\usage{
extract_my_zips(dirpath, all = FALSE, recurse = FALSE)
}
\arguments{
\item{dirpath}{directory path}

\item{all}{logical (default = FALSE). All posible archive formats or just .zip?}

\item{recurse}{logical (default = FALSE). as used in {fs} pkg functions.
If TRUE, recurse fully. If a positive number, the number of levels to recurse.}
}
\value{
a list
}
\description{
Used to be called sort_my_zips
}
\details{
Sometimes I end up accumilating a large collection of zip files
and I just want to quickly extract all of them, so (naturally)
I automated this process in R.
}
\examples{
\dontrun{
out <- fs::path_wd() \%>\% extract_my_zips()
out$zips_to_check #fix these
out$problematic_extractions$name #check these
fs::file_delete(out$my_zips$path) #clean Up
}

}
